Change RegisterSerializer implementation -The gender field in RegisterSerializer is now a serializer ChoiceField -The validate_gender method in RegiterSerializer was eliminated -Tests.py was deleted -Now there is a tests folder wich contains test_edit_profile.py and test_signup.py
